pyproject.toml
[tool.poetry]
name = "trustpoint"
version = "0.1.1"
description = "Trustpoint Server Software"
authors = ["TrustPoint-Project"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyasn1 = "^0.6.1"
pyasn1-modules = "^0.4.1"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
django-ninja = "^1.3.0"
croniter = "^5.0.1"
django = "^5.1.3"
cryptography = "^43.0.3"
django-taggit = "^6.1.0"
django-filter = "^24.3"
zeroconf = "^0.136.0"
django-tables2 = "^2.7.0"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"
sphinxcontrib-plantuml = "^0.30"
devtools = "^0.12.2"
django-extensions = "^3.2.3"
pyopenssl = "^24.2.1"
werkzeug = "^3.1.3"
furo = "^2024.8.6"
ruff = "^0.7.3"
pillow = "^11.0.0"
docutils-stubs = "^0.0.22"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4


[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN101", "ANN102", "ANN401",
    "D203", "D213", "D415",
    "TD003",
    # TODO(Alex): FIX should be added again, after we are departing from the PoC phase
    "FIX",
    "COM812",               # Not compatible with the formatter
    "ISC001"                # Not compatible with the formatter
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"